superadmin permissions added to questions controller

+++ app/controllers/questions_controller.rb
@@ -4,44 +4,72 @@ class QuestionsController < ApplicationController
   end
 
   def new
-    @question = Question.new
-  end
-
-  def create
-    @question = Question.new(question_params) #this will generate a question object
-    # binding.pry
-    if @question.save
-      redirect_to questions_path
+    @admin = Admin.find_by(id: current_admin.id)
+    if @admin.super_admin == true
+      @question = Question.new
     else
-      #render form again
+      redirect_to "/questions"
     end
   end
 
+  def create
+    @admin = Admin.find_by(id: current_admin.id)
+    if @admin.super_admin == true
+      @question = Question.new(question_params) #this will generate a question object
+      if @question.save
+        redirect_to questions_path
+      else
+        #render form again
+      end
+    else
+      redirect_to "/questions"
+    end
+  end
+
+  def show
+    redirect_to "/questions"
+  end
+
   def edit
-    @question = Question.where(id: params[:id]).first
+    @admin = Admin.find_by(id: current_admin.id)
+    if @admin.super_admin == true
+      @question = Question.where(id: params[:id]).first
+    else
+      redirect_to "/questions"
+    end
   end
 
   def update
-    @question = Question.where(id: params[:id]).first
-    if @question.update_attributes(question_params)
-      redirect_to questions_path
+    @admin = Admin.find_by(id: current_admin.id)
+    if @admin.super_admin == true
+      @question = Question.where(id: params[:id]).first
+      if @question.update_attributes(question_params)
+        redirect_to questions_path
+      else
+        # render the edit form again
+      end
     else
-      # render the edit form again
+      redirect_to "/questions"
     end
   end
 
   def destroy
-    @question = Question.where(id: params[:id]).first
-    if @question.destroy
-      redirect_to questions_path
-    else
+    @admin = Admin.find_by(id: current_admin.id)
+    if @admin.super_admin == true
+      @question = Question.where(id: params[:id]).first
+      if @question.destroy
+        redirect_to questions_path
+      else
 
+      end
+    else
+      redirect_to "/questions"
     end
-  end 
+  end
 
   private
 
   def question_params
-    params.require(:question).permit(:question, :answer)
+    params.require(:question).permit(:question_text, :answer)
   end
 end
